[Justice Counts] Bulk Publishing: Generate and send list of unchanged reports after spreadsheet upload to the FE (Recidiviz/recidiviz-data#20534)
## Description of the change

There is one edge-case surrounding a user uploading a spreadsheet that
is intended to update existing reports but makes no changes to them
(because the values are the same). Currently, the BE only has references
to:
* existing reports that have been updated based on changes to the values
detected in the spreadsheet ingest process
* newly created reports as a result of the spreadsheet ingest process

Since there's no reference to the unchanged reports - when you hit this
edge case, you are unable to proceed with publishing because there's no
list of reports to pass to the publish endpoint - and thus, you're met
with an error.

This PR extracts a list of unchanged reports after ingesting a
spreadsheet in order to:
* render a list of unchanged reports in the Data Upload -> Publish
Review page in the FE
* allow a user to publish those unchanged reports from the review page

The FE companion PR can be found here:
https://github.com/Recidiviz/justice-counts/pull/596

## Related issues

Closes Recidiviz/recidiviz-data#20533

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Mahmoud <mahmoud@Mahmouds-MBP.cable.rcn.com>
GitOrigin-RevId: 28ef7952479f87567b435b7720a343779f3ed190