[US_MI][Workflows] Add recommended supervision level to supervision level downgrade opportunity record query (Recidiviz/recidiviz-data#20553)
## Description of the change

Here I add recommended supervision level as a case notes entry to the
opportunity record query.

It is important to note that Medium w/ Override to Max levels
('MEDIUM/HIGH', 'HIGH/MEDIUM', 'HIGH/LOW', 'LOW/HIGH') here are
recommended for downgrade from Intensive --> MAXIMUM whereas for the
classification review form, where we are recommending supervision level
based off of COMPAS scores only in very specific cases we recommend
these levels for MEDIUM.

We recommend MAXIMUM in supervision level downgrade as this opportunity
is meant to surface mismatches of supervision level and assessment level
in the first 4-12 months of supervision before the first classification
review. If a client has Medium w/ Override to Max and is on `MAXIMUM` it
is often the case that those levels are NOT a mismatch but an Override,
so we are only surfacing these clients if they have a `HIGH` supervision
level.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 114be1732e01fe58304f89a317b4d0f3bad47d88