[US_PA] Update supervision contact external ID (Recidiviz/recidiviz-data#20589)
## Description of the change

Ingest is blocked during a partial rerun of supervision contacts in PA
staging (Recidiviz/recidiviz-data#20568) by the following error:
```
Found multiple different ingested entities of type [StateSupervisionContact] with conflicting information: ['StateSupervisionContact(external_id=<redacted>, supervision_contact_id=None)', 'StateSupervisionContact(external_id=<redacted, identical to previous>, supervision_contact_id=None)']"
```
StateSupervisionContact `external_id` fields were constructed by
concatenating four fields:
- parole_number
- created_date
- contact_type
- duration

This approach was sufficient when PKs for the raw data were insufficient
and used the same set of fields. After Recidiviz/recidiviz-data#20545 updated the PKs in that
raw data, it became clear that these four fields are identical across
some distinct supervision contacts. That's what's causing this error.

This change updates the `external_id` of StateSupervisionContact in the
`supervision_contacts` ingest mapping to be the same as the primary keys
in the raw data: the unique ID for a given supervision contact, and the
unique ID of the person on supervision related to that contact. It also
removes the now-unused `duration` and `created_date` fields from the
ingest view, and adds an ingest view test for this view (Recidiviz/recidiviz-data#18848).

Once this change is deployed, ingest view results will need to be
cleared again for `state_supervision_contact` in PA staging primary and
regenerated with this new set of fields, which should resolve the entity
matching error.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: db4b7252075d011619f2d03a693354147eaa9cf4