[US_IX] Ingest StateEarlyDischarge (Recidiviz/recidiviz-data#20119)
## Description of the change

Ingest StateEarlyDischarge for US_IX, which is done in three separate
views:
- `early_discharge_parole`: Ingest all parole early discharge requests
using Atlas data
- `early_discharge_probation_legacy`: Ingest probation early discharge
requests pre-Atlas migration using legacy Idaho data
- Based on the US_ID
[view](https://github.com/Recidiviz/recidiviz-data/blob/aff1af1b1828bc6e6b57926d690b6186aa4a52fa/recidiviz/ingest/direct/regions/us_id/ingest_views/view_early_discharge_supervision_sentence.py)
with some modifications. In US_ID, there was also a view to deal with
the ["deleted"/invalid early discharge
requests](https://github.com/Recidiviz/recidiviz-data/issues/3746), but
since legacy ID data will never be updated and we'll just use the last
full historical copy of this data, I don't think we need to deal with
that edge case for US_IX.
- `early_discharge_probation`: Ingest probation early discharge requests
post-Atlas migration using Atlas data

In Atlas, data on early discharge requests are recorded at the
individual, not at the sentence level. Though the legacy data we're
using to ingest legacy probation early discharge requests is on the
sentence level, there's no way to link sentences in the legacy data to
sentences in Atlas. Therefore, for each `StateEarlyDischarge` entity
that we're ingesting, instead of hanging it off an existing US_IX
sentence, we'll create a dummy sentence to hang it off of. Have
confirmed with @samanthahuff and @mayukas that using a dummy sentence
will not impact things downstream for impact tracking or outliers.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a4b34a1a840f052a0be57af490faf639fbb422f0