[Admin Panel] Fix freshness page (Recidiviz/recidiviz-data#20699)
## Description of the change

In Recidiviz/recidiviz-data#20104 I updated our usage of `CloudSqlToBqRefreshStatus` to store a
timezone-aware datetime. We use this datetime as a literal in BigQuery
views to power the freshness page, and BigQuery only allows
timezone-naive datetimes. This change converts the datetime back to
timezone-naive before adding it to the query.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

closes Recidiviz/recidiviz-data#20413

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 20c0521e98652af2b677c4f21fb4a08cf5f2dc9d