[US_MI] Fix projected_min_release_date (Recidiviz/recidiviz-data#20687)
## Description of the change

A few months ago, I put in a
[PR](https://github.com/Recidiviz/recidiviz-data/pull/18360) to revise
the mappings for incarceration sentences because I was incorrectly
mapping `expiration_date` to `projected_min_release_date` when I should
have been mapping it to `projected_max_release_date`. Basically, the
revision changed `projected_min_release_date` so that we're no longer
hydrating it. However, when I did a partial rerun for that PR (since it
included other changes as well), I must have not cleared out the state
database correctly because `projected_min_release_date` is still
currently hydrated for incarceration sentences in MI in Postgres and BQ.

This PR cleans up ^ this mess AND revises `projected_min_release_date`
to be set equal to the date we calculate for `parole_eligibility_date`.
So this PR includes the mapping change for the correct hydration of
`projected_min_release_date` as well as a database migration to set
`projected_min_release_date=parole_eligibility_date`. In that database
migration, the downgrade migration is to set
`projected_min_release_date=NULL` since that's what it should have been
had I cleared out the state database correctly.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e15fd41b2339ff452e5923c053c4acf1d8a8d94f