[US_MI] Combine VFO and non-VFO scales into one in ingest (Recidiviz/recidiviz-data#19595)
## Description of the change

This PR:
- Combines VFO and non-VFO scales into one assessment row in ingest so
that it doesn't have to be done in assessments pre-processing (see
Recidiviz/recidiviz-data#19465). For the raw score, we take the maximum score between the two
scales, though we don't actually currently use raw score anywhere.
- Uses `FKCoSyRecSupervisionLevel` (aka recommended supervision level)
to map `AssessmentLevel` when the `ScoreText` value is inconclusive.
- Adds in `recommended_level` into the `assessments_metadata`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 15fcc1f86f7c262eec5d977493e8bd98d7caaa74