[Justice Counts] changing 'Category Not Recognized' error to a warning (Recidiviz/recidiviz-data#20676)
## Description of the change

Previously, if a user uploaded a file via bulk upload that contained a
deprecated dimensions, an error was raised in
bulk_upload_helpers.fuzzy_match_against_options(). This caused the
ingest of the rest of the data in that time range to stop. In order to
continue ingest of other data in the same time range, we are appending a
warning to metric_key_to_errors, rather than raising an error. We now
see the appropriate title and warnings on the bulk upload errors page:
<img width="877" alt="Screenshot 2023-05-04 at 11 50 54 AM"
src="https://user-images.githubusercontent.com/82831800/236261543-43875e7c-3ec1-4c9a-ab8a-c032c77071f6.png">

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@Michelles-MacBook-Pro.local>
GitOrigin-RevId: 807d39419e08422561de6293c28a52637f0edf70