[Justice Counts] Recurring Reports fiscal year fix (Recidiviz/recidiviz-data#20825)
## Description of the change

Lili noticed that all agencies in production had Fiscal Year (May)
reports created, even though these agencies did not have any fiscal
metrics reported on the May month timeline. This occurred because
DatapointsForMetric.get_metric_definitions_for_report() returns an empty
list of metric_definitions [], however we were only checking if
annual_metric_defs was None (and not an empty list).

This bug should now be fixed as we are checking that annual_metric_defs
is not None and the list of annual_metric_defs is empty.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#20765

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: Michelle Orden <michelleorden@michelles-mbp.lan>
GitOrigin-RevId: ae149598b844f07701bef0360ac23137a1318f9f